convert-all-flts: Stop monitoring, run flt2vhs with >1 .flt
flt2vhs now takes many flt files at once to merge them,
which means that we can't (or at least shouldn't) call it on one file at
a time as BMS finishes writing them. Rely on the no-op patcher to keep
BMS from doing silly things, and make this a simple tool that renames
everything, converts it, and gets out.